requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.5.2
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
dearpygui==1.8.0
distlib==0.3.6
Django==4.1.3
django-debug-toolbar==4.2.0
django-filter==23.5
django-rest-swagger==2.2.0
djangorestframework==3.14.0
exceptiongroup==1.2.0
fastapi==0.109.0
filelock==3.8.0
fonttools==4.39.4
frozenlist==1.4.1
fsspec==2023.12.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-metadata==7.0.1
importlib-resources==5.12.0
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.4
llvmlite==0.41.1
magic-filter==1.0.12
Mako==1.3.0
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.7.1
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numba==0.58.1
numpy==1.24.3
openai-whisper==20231117
openapi-codec==1.3.2
packaging==23.1
Pillow==9.5.0
platformdirs==2.5.4
psycopg2==2.9.9
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
pygame==2.1.2
pyparsing==3.0.9
PyQt6==6.4.2
PyQt6-Qt6==6.4.3
PyQt6-sip==13.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
regex==2023.12.25
requests==2.31.0
scipy==1.10.1
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
SpeechRecognition==3.8.1
starlette==0.35.1
sympy==1.12
tiktoken==0.5.2
tk==0.1.0
torch==2.1.2
tqdm==4.66.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.26.0
virtualenv==20.16.7
yarl==1.9.4
zipp==3.15.0
